feat: split switch into a control and field composables
