+++ packages/core/src/useSwitch/index.ts
@@ -1,1 +1,2 @@
 export * from './useSwitch';
+export * from './useSwitchControl';
